Break out event emitting into its own class so the instance can be persistent
