+++ src/device-listener/index.ts
@@ -5,6 +5,7 @@ SPDX-License-Identifier: AGPL-3.0-only OR GPL-3.0-only OR LicenseRef-Element-Com
 Please see LICENSE files in the repository root for full details.
 */
 
+export * from "./CurrentDeviceChangedEmitter";
 export * from "./DeviceListener";
 export * from "./DeviceListenerCurrentDevice";
 export * from "./DeviceListenerOtherDevices";
